fix: use-leave-page hook example
