+++ packages/ui/src/styles/globals.css
@@ -8,6 +8,7 @@
 @custom-variant dark (&:is(.dark *));
 
 :root {
+  --color-fd-primary: oklch(0.65 0.18 132);
   --background: oklch(0.9979 0.0066 115.7004);
   --foreground: oklch(0.423 0.0426 123.7122);
   --card: oklch(1 0 0);
@@ -60,6 +61,7 @@
 }
 
 .dark {
+  --color-fd-primary: oklch(0.8972 0.1786 126.6652);
   --background: oklch(0.1591 0 0);
   --foreground: oklch(0.8871 0.0136 115.8605);
   --card: oklch(0.1953 0.002 106.5807);
